requirements.txt
aioredis==1.3.0
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.2
channels==2.1.2
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
dj-database-url==0.5.0
Django==2.2.7
django-configurations==2.1
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-extensions==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
whitenoise==4.1.4
zope.interface==4.7.1
